+++ omnijson/omni_getbalance.go
@@ -16,7 +16,7 @@ type OmniGetBalanceResult = struct {
 
 type OmniGetBalanceCommand struct {
 	Address    string
-	PropertyID int32
+	PropertyID int64
 }
 
 func (OmniGetBalanceCommand) Method() string {
